refactor: split domain-tools.spec.ts into per-domain test files
Separate schedule, secret, and human-input tests into individual files
for better organization and discoverability.
